+++ packages/harpy-cli/templates/app/src/app.module.ts
@@ -1,11 +1,20 @@
 import { Module } from '@nestjs/common';
+import { I18nModule } from '@hepta-solutions/harpy-core';
 import { HomeModule } from './features/home/home.module';
 import { AboutModule } from './features/about/about.module';
 import { AuthModule } from './features/auth/auth.module';
 import { DashboardModule } from './features/dashboard/dashboard.module';
+import { i18nConfig } from './i18n/i18n.config';
 
 @Module({
-  imports: [HomeModule, AboutModule, AuthModule, DashboardModule],
+  imports: [
+    // Configure I18n module
+    I18nModule.forRoot(i18nConfig),
+    HomeModule,
+    AboutModule,
+    AuthModule,
+    DashboardModule,
+  ],
   controllers: [],
   providers: [],
 })